Enhance auto-response detection and categorization
This commit improves the handling of automated emails and
system-generated messages:

- Renames `01-autoresponders.sieve` to `01-autoreplies.sieve` and
enhances its logic to better detect auto-responses based on common
headers like `X-Auto-Response-Suppress`, `Precedence`, and
`Auto-Submitted`. It also files these into "Not-a-human" and "Activity
Stream" labels.

- Adds `01-no-return-path.sieve` to label emails without a return path
as "Not-a-human", as these are typically system-generated.

- Updates `05-action-emails.sieve` to use the "Not-a-human" label
instead of "Autoresponse" and adjusts retention periods.

- Adds `06-caa-domain.sieve` and `06-dmarc-reports.sieve` to handle
emails from specific domains related to domain authentication and file
them into "Service Events/OPS" and "Not-a-human" with a 90-day
retention.

- Renames `02-deliverability-events.sieve` to `06-deliverability.sieve`
and updates it to use the "Not-a-human" label.

- Adds `10-promotional.sieve` to identify and handle promotional/bulk
emails based on the `Precedence` header.

Other changes:

- Updates retention periods for some rules in `15-activity-github.sieve`
and `15-activity-mailinglists.sieve`.
- Removes an unused filter from `00-testing.sieve`.
- Adds a link to a Sieve syntax checker in the README.